Audit cleanup: dedupe patchwork, unify conventions across screens
Resolves an audit of inconsistent/duplicated implementations that
accumulated across screens. No user-facing behavior changes except where
noted (confirm dialogs, onboarding landscape, shading default).

- Backup: round-trip QuickCapture, Reminder, and Update prefs (were silently dropped)
- Theme: collapse surface shading to a single stored key (default = medium); drop legacy enum/bool keys
- Layout: add shared isLandscape()/isSmallLandscape() helpers; route all screens through them and fix onboarding using a different measurement
- Editor: extract shared BaseEditorViewModel from the near-identical note/list edit screens
- Dialogs: standardize all confirmations on AlertDialog + RememberTextButton (Trash/Archive confirms were bottom sheets)
- Empty states: share one spacing helper between home and history
- Workers: align trash-sweep scheduling policy with backup (UPDATE)
- Import: use injected @IoDispatcher instead of hardcoded Dispatchers.IO
- Help: expose read-only state flows and match the app's stateIn policy
- Dates: move note-card formatting to java.time; document backup filename's intentional fixed format
- Docs: comment why the update sheet keeps a raw ModalBottomSheet (landscape drag-to-dismiss)
