+++ app/src/main/java/dev/bikram/remember/data/ReminderPrefs.kt
@@ -7,6 +7,7 @@ import androidx.datastore.preferences.preferencesDataStore
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
+import org.json.JSONObject
 
 data class ReminderPreferencesState(
     val keepReminderNotificationsUntilDone: Boolean = false,
@@ -42,6 +43,26 @@ class ReminderPrefs(
     suspend fun setReminderSummaryNotificationEnabled(enabled: Boolean) {
         context.reminderDataStore.edit { prefs ->
             prefs[Keys.SUMMARY_NOTIFICATION] = enabled
+        }
+    }
+
+    suspend fun exportForBackup(): JSONObject {
+        val prefs = context.reminderDataStore.data.first()
+        return JSONObject().apply {
+            put(Keys.KEEP_UNTIL_DONE.name, prefs[Keys.KEEP_UNTIL_DONE] ?: false)
+            put(Keys.SUMMARY_NOTIFICATION.name, prefs[Keys.SUMMARY_NOTIFICATION] ?: false)
+        }
+    }
+
+    suspend fun importFromBackup(json: JSONObject?) {
+        if (json == null || json.length() == 0) return
+        context.reminderDataStore.edit { mutable ->
+            if (json.has(Keys.KEEP_UNTIL_DONE.name) && !json.isNull(Keys.KEEP_UNTIL_DONE.name)) {
+                mutable[Keys.KEEP_UNTIL_DONE] = json.getBoolean(Keys.KEEP_UNTIL_DONE.name)
+            }
+            if (json.has(Keys.SUMMARY_NOTIFICATION.name) && !json.isNull(Keys.SUMMARY_NOTIFICATION.name)) {
+                mutable[Keys.SUMMARY_NOTIFICATION] = json.getBoolean(Keys.SUMMARY_NOTIFICATION.name)
+            }
         }
     }
