+++ app/src/main/java/dev/bikram/remember/ui/components/AlertChrome.kt
@@ -45,7 +45,6 @@ import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.layout.boundsInWindow
 import androidx.compose.ui.layout.layout
 import androidx.compose.ui.layout.onPlaced
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.LocalWindowInfo
 import androidx.compose.ui.res.pluralStringResource
@@ -59,6 +58,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.isSpecified
 import androidx.graphics.shapes.Morph
 import dev.bikram.remember.R
+import dev.bikram.remember.ui.common.isSmallLandscape
 import dev.bikram.remember.ui.common.RememberMaterialRoundedSymbol
 import dev.bikram.remember.ui.theme.MorphPolygonShape
 import dev.bikram.remember.ui.theme.reducedMotionAwareSpec
@@ -100,9 +100,7 @@ fun AlertFloatingFab(
 ) {
     if (summary.count <= 0) return
 
-    val configuration = LocalConfiguration.current
-    val isLandscape = configuration.orientation == android.content.res.Configuration.ORIENTATION_LANDSCAPE
-    val isSmallLandscape = isLandscape && configuration.screenHeightDp < 480
+    val isSmallLandscape = isSmallLandscape()
 
     val label = stringResource(R.string.main_alert_fab_label)
     val scheme = MaterialTheme.colorScheme
@@ -268,9 +266,7 @@ fun AlertFloatingActionButtonMenu(
     // must extend left of the FAB, which is impossible inside that clip. The FAB anchors
     // a plain Box instead, and the bars render as an unclipped sibling placed above it.
     BackHandler(enabled = expanded) { onExpandedChange(false) }
-    val configuration = LocalConfiguration.current
-    val isLandscape = configuration.orientation == android.content.res.Configuration.ORIENTATION_LANDSCAPE
-    val isSmallLandscape = isLandscape && configuration.screenHeightDp < 480
+    val isSmallLandscape = isSmallLandscape()
     val barIconSize = if (isSmallLandscape) rememberResponsiveActionButtonSize() else 44.dp
     val barContentScale = if (isSmallLandscape) barIconSize.value / 44f else 1f
     val density = LocalDensity.current
