+++ app/src/main/java/dev/bikram/remember/ui/edit/EditNoteViewModel.kt
@@ -1,159 +1,36 @@
 package dev.bikram.remember.ui.edit
 
-import android.net.Uri
 import androidx.lifecycle.SavedStateHandle
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import dagger.hilt.android.EntryPointAccessors
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.bikram.remember.data.AppMediaStorage
-import dev.bikram.remember.data.Importance
-import dev.bikram.remember.data.NoteAction
-import dev.bikram.remember.data.NoteAttachmentEntity
 import dev.bikram.remember.data.NoteOptions
 import dev.bikram.remember.data.NoteRepository
-import dev.bikram.remember.data.RecurrenceRule
-import dev.bikram.remember.data.RememberReservedTags
-import dev.bikram.remember.di.SettingsDependenciesEntryPoint
-import dev.bikram.remember.ui.common.HeroFraming
-import dev.bikram.remember.ui.nav.Routes
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import javax.inject.Inject
-import dev.bikram.remember.data.Visibility as NoteVisibility
 
 /**
- * Owns every persisted field for the Edit Note screen and serializes save/load through
- * a single [EditorPersistenceSession] so that ON_STOP, dispose, and explicit calls cannot race against
- * each other (or against the async note load in [init]).
- *
- * Each field is exposed as its own [StateFlow] so leaf composables can collect just the
- * slice they render and skip recompositions when unrelated fields change. The screen-level
- * composable should not collect everything in one place.
+ * Owns the free-text body for the Edit Note screen on top of every persisted field shared with
+ * lists in [BaseEditorViewModel]. Only the body payload and the create/update/diff calls that
+ * depend on it live here; setters, tag handling, and the note-lifecycle actions are inherited.
  */
 @HiltViewModel
 class EditNoteViewModel
     @Inject
     constructor(
-        private val repository: NoteRepository,
-        private val appMediaStorage: AppMediaStorage? = null,
+        repository: NoteRepository,
+        appMediaStorage: AppMediaStorage? = null,
         savedStateHandle: SavedStateHandle,
-    ) : ViewModel() {
-        private val noteId: Long? =
-            savedStateHandle
-                .get<Long>(Routes.ARG_ID)
-                ?.takeIf { it > 0L }
-        private val prefillBody: String = savedStateHandle[Routes.ARG_PREFILL] ?: ""
-
-        private val _title = MutableStateFlow("")
-        val title: StateFlow<String> = _title.asStateFlow()
-
+    ) : BaseEditorViewModel(repository, appMediaStorage, savedStateHandle) {
         private val _body = MutableStateFlow(if (noteId == null) prefillBody else "")
         val body: StateFlow<String> = _body.asStateFlow()
-
-        private val _starred = MutableStateFlow(false)
-        val starred: StateFlow<Boolean> = _starred.asStateFlow()
-
-        /**
-         * Snapshot of the underlying note's [dev.bikram.remember.data.NoteEntity.completedAt]
-         * as a boolean. Driven by the live DB observer in [init], so external completion
-         * via swipe / notification action / repository.markCompleted is reflected here too.
-         */
-        private val _completed = MutableStateFlow(false)
-        val completed: StateFlow<Boolean> = _completed.asStateFlow()
-
-        private val _reminderAt = MutableStateFlow<Long?>(null)
-        val reminderAt: StateFlow<Long?> = _reminderAt.asStateFlow()
-
-        private val _recurrence = MutableStateFlow<RecurrenceRule?>(null)
-        val recurrence: StateFlow<RecurrenceRule?> = _recurrence.asStateFlow()
-
-        private val _importance = MutableStateFlow(Importance.DEFAULT)
-        val importance: StateFlow<Importance> = _importance.asStateFlow()
-
-        private val _visibility = MutableStateFlow(NoteVisibility.DEFAULT)
-        val visibility: StateFlow<NoteVisibility> = _visibility.asStateFlow()
-
-        private val _locked = MutableStateFlow(false)
-        val locked: StateFlow<Boolean> = _locked.asStateFlow()
-
-        private val _pictureUri = MutableStateFlow<String?>(null)
-        val pictureUri: StateFlow<String?> = _pictureUri.asStateFlow()
-
-        /** Bumped whenever the hero bytes change, including in-place edits that keep the same URI string. */
-        private val _pictureRevision = MutableStateFlow(0L)
-        val pictureRevision: StateFlow<Long> = _pictureRevision.asStateFlow()
-
-        private val _pictureHeroFraming = MutableStateFlow<String?>(null)
-        val pictureHeroFraming: StateFlow<String?> = _pictureHeroFraming.asStateFlow()
-
-        private val _iconKey = MutableStateFlow<String?>(null)
-        val iconKey: StateFlow<String?> = _iconKey.asStateFlow()
-
-        private val _actions = MutableStateFlow<List<NoteAction>>(emptyList())
-        val actions: StateFlow<List<NoteAction>> = _actions.asStateFlow()
-
-        private val _tags = MutableStateFlow<List<String>>(emptyList())
-        val tags: StateFlow<List<String>> = _tags.asStateFlow()
-        val activeTagSuggestions: StateFlow<List<String>> =
-            repository
-                .observeActiveTagSuggestions()
-                .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5_000), emptyList())
-
-        private val _attachments = MutableStateFlow<List<NoteAttachmentEntity>>(emptyList())
-        val attachments: StateFlow<List<NoteAttachmentEntity>> = _attachments.asStateFlow()
-
-        private val _createdAt = MutableStateFlow<Long?>(null)
-        val createdAt: StateFlow<Long?> = _createdAt.asStateFlow()
-
-        private val _updatedAt = MutableStateFlow<Long?>(null)
-        val updatedAt: StateFlow<Long?> = _updatedAt.asStateFlow()
-
-        /**
-         * Mirrors the underlying note's archived / trashed shelf. Used by the edit screen to flip
-         * into read-only mode and swap the bottom-bar action set. New notes always start on the
-         * active shelf.
-         */
-        private val _archived = MutableStateFlow(false)
-        val archived: StateFlow<Boolean> = _archived.asStateFlow()
-
-        private val _trashed = MutableStateFlow(false)
-        val trashed: StateFlow<Boolean> = _trashed.asStateFlow()
-
-        /**
-         * True once this session is backed by a database row ([loadedId] non-null). New drafts start
-         * false so the action bar omits archive/trash until the first save (or attachment create)
-         * assigns an id, without tying that UI to the navigation argument (which stays null).
-         */
-        private val _hasPersistedRow = MutableStateFlow(noteId != null)
-        val hasPersistedRow: StateFlow<Boolean> = _hasPersistedRow.asStateFlow()
-
-        private val _currentNoteId = MutableStateFlow(noteId)
-        val currentNoteId: StateFlow<Long?> = _currentNoteId.asStateFlow()
 
         /** True after the initial DB load has populated the state flows (or immediately for a new note). */
         private val _loaded = MutableStateFlow(noteId == null)
         val loaded: StateFlow<Boolean> = _loaded.asStateFlow()
-
-        private val persistence = EditorPersistenceSession(initialDirty = noteId == null && prefillBody.isNotBlank())
-        val hasUnsavedChanges: StateFlow<Boolean> = persistence.hasUnsavedChanges
-
-        private var loadedId: Long? = noteId
-
-        private var originalNote: dev.bikram.remember.data.NoteEntity? = null
-
-        private fun markDirty() {
-            persistence.markDirty()
-        }
-
-        private fun syncHasPersistedRow() {
-            _hasPersistedRow.value = loadedId != null
-            _currentNoteId.value = loadedId
-        }
 
         init {
             if (noteId != null) {
@@ -162,27 +39,8 @@ class EditNoteViewModel
                     try {
                         val existing = repository.get(noteId)
                         if (existing != null) {
-                            val n = existing.note
-                            originalNote = n
-                            _title.value = n.title
-                            _body.value = n.body
-                            _starred.value = n.starred || n.tags.contains(RememberReservedTags.STARRED)
-                            _reminderAt.value = n.reminderAt
-                            _recurrence.value = n.recurrence?.sanitized()
-                            _importance.value = n.importance
-                            _visibility.value = n.visibility
-                            _locked.value = n.locked
-                            _pictureUri.value = n.pictureUri
-                            _pictureHeroFraming.value = n.pictureHeroFraming
-                            _iconKey.value = n.iconKey
-                            _actions.value = n.actions
-                            _tags.value = n.tags.filterNot { it == RememberReservedTags.STARRED }
-                            _attachments.value = existing.attachments
-                            _archived.value = n.archived
-                            _trashed.value = n.trashed
-                            _completed.value = n.completedAt != null
-                            _createdAt.value = n.createdAt
-                            _updatedAt.value = n.updatedAt
+                            applyLoadedCommon(existing)
+                            _body.value = existing.note.body
                         }
                     } finally {
                         // Leave loading when the load finishes: missing row, success, or thrown from get().
@@ -190,33 +48,8 @@ class EditNoteViewModel
                         persistence.unlock()
                     }
                 }
-                // Live-mirror only the fields that can be written from OUTSIDE this VM
-                // while the editor is open: reminderAt + recurrence (snooze action,
-                // recurrence advance after a fire, mark-as-done from notification). The
-                // other fields are owned by this VM's user-input flow and would clobber
-                // unsaved drafts if we mirrored them too. Room's Flow only emits on
-                // actual row changes, so this is a single small distinct subscription -
-                // negligible battery impact, no polling.
-                viewModelScope.launch {
-                    repository.observe(noteId).collect { row ->
-                        val n = row?.note ?: return@collect
-                        if (_reminderAt.value != n.reminderAt) _reminderAt.value = n.reminderAt
-                        val sanitized = n.recurrence?.sanitized()
-                        if (_recurrence.value != sanitized) _recurrence.value = sanitized
-                        if (_trashed.value != n.trashed) _trashed.value = n.trashed
-                        val isCompleted = n.completedAt != null
-                        if (_completed.value != isCompleted) _completed.value = isCompleted
-                        if (_createdAt.value != n.createdAt) _createdAt.value = n.createdAt
-                        if (_updatedAt.value != n.updatedAt) _updatedAt.value = n.updatedAt
-                    }
-                }
+                startExternalFieldMirror(noteId)
             }
-        }
-
-        fun setTitle(value: String) {
-            if (_title.value == value) return
-            _title.value = value
-            markDirty()
         }
 
         fun setBody(value: String) {
@@ -225,215 +58,19 @@ class EditNoteViewModel
             markDirty()
         }
 
-        fun toggleStar() {
-            _starred.value = !_starred.value
-            markDirty()
+        override suspend fun persistNewDraftForAttachment(): Long {
+            val newId =
+                repository.createNote(
+                    title = _title.value,
+                    body = _body.value,
+                    colorIndex = 0,
+                    options = currentOptions(),
+                )
+            loadedId = newId
+            syncHasPersistedRow()
+            if (_starred.value) repository.setStarred(newId, true)
+            return newId
         }
-
-        fun setReminder(
-            at: Long?,
-            rule: RecurrenceRule?,
-        ) {
-            val normalized = rule?.sanitized()
-            if (_reminderAt.value == at && _recurrence.value == normalized) return
-            _reminderAt.value = at
-            _recurrence.value = normalized
-            markDirty()
-        }
-
-        fun setImportance(value: Importance) {
-            if (_importance.value == value) return
-            _importance.value = value
-            markDirty()
-        }
-
-        fun setVisibility(value: NoteVisibility) {
-            if (_visibility.value == value) return
-            _visibility.value = value
-            markDirty()
-            refreshActiveNotificationVisibility(value)
-        }
-
-        private fun refreshActiveNotificationVisibility(value: NoteVisibility) {
-            val id = loadedId ?: return
-            viewModelScope.launch {
-                repository.refreshNotificationVisibilityPreview(id, value)
-            }
-        }
-
-        fun toggleLock() {
-            _locked.value = !_locked.value
-            markDirty()
-        }
-
-        fun setPictureUri(value: String?) {
-            if (_pictureUri.value != value) {
-                _pictureUri.value = value
-            }
-            if (value == null) {
-                _pictureHeroFraming.value = null
-            }
-            _pictureRevision.value = _pictureRevision.value + 1L
-            markDirty()
-        }
-
-        fun setHeroWithFraming(
-            pictureUri: String,
-            framing: HeroFraming,
-        ) {
-            _pictureUri.value = pictureUri
-            _pictureHeroFraming.value = framing.toJsonString()
-            _pictureRevision.value = _pictureRevision.value + 1L
-            markDirty()
-        }
-
-        fun setIconKey(value: String?) {
-            if (_iconKey.value == value) return
-            _iconKey.value = value
-            markDirty()
-        }
-
-        fun setActions(value: List<NoteAction>) {
-            if (_actions.value == value) return
-            _actions.value = value
-            markDirty()
-        }
-
-        fun setTags(value: List<String>) {
-            val cleaned = value.filterNot { it == RememberReservedTags.STARRED }
-            if (_tags.value == cleaned) return
-            _tags.value = cleaned
-            markDirty()
-        }
-
-        fun addTag(
-            value: String,
-            colorHex: String,
-        ) {
-            val cleaned = value.trim()
-            if (cleaned.isBlank() || cleaned == RememberReservedTags.STARRED) return
-            if (_tags.value.any { tag -> tag.equals(cleaned, ignoreCase = true) }) return
-            _tags.value = _tags.value + cleaned
-            markDirty()
-            viewModelScope.launch {
-                repository.tagRepository?.setTagColor(cleaned, colorHex)
-            }
-        }
-
-        fun saveTagsWithColors(
-            tags: List<String>,
-            newColors: Map<String, String>,
-        ) {
-            setTags(tags)
-            if (newColors.isNotEmpty()) {
-                viewModelScope.launch {
-                    newColors.forEach { (name, hex) ->
-                        repository.tagRepository?.setTagColor(name, hex)
-                    }
-                }
-            }
-        }
-
-        fun editExistingTag(
-            oldName: String,
-            newName: String,
-            colorHex: String?,
-            resetColor: Boolean,
-        ) {
-            viewModelScope.launch {
-                val result =
-                    repository.tagRepository?.editTag(
-                        oldName = oldName,
-                        newName = newName,
-                        colorHex = colorHex,
-                        resetColor = resetColor,
-                    ) ?: return@launch
-                val updatedTags =
-                    _tags.value
-                        .map { tagName ->
-                            if (tagName.equals(result.oldName, ignoreCase = true)) result.newName else tagName
-                        }.distinctBy { tagName -> tagName.lowercase() }
-                if (_tags.value != updatedTags) {
-                    _tags.value = updatedTags
-                    markDirty()
-                }
-            }
-        }
-
-        fun addAttachment(
-            uri: Uri,
-            name: String,
-            mime: String?,
-        ) {
-            viewModelScope.launch {
-                persistence.withLock {
-                    val id =
-                        loadedId ?: run {
-                            val newId =
-                                repository.createNote(
-                                    title = _title.value,
-                                    body = _body.value,
-                                    colorIndex = 0,
-                                    options = currentOptions(),
-                                )
-                            loadedId = newId
-                            syncHasPersistedRow()
-                            if (_starred.value) repository.setStarred(newId, true)
-                            newId
-                        }
-                    val attachmentUri =
-                        appMediaStorage
-                            ?.copyAttachmentToPrivateStorage(
-                                noteId = id,
-                                sourceUri = uri,
-                                displayName = name,
-                                mimeType = mime,
-                            )?.uriString ?: uri.toString()
-                    repository.addAttachment(id, attachmentUri, name, mime)
-                    _attachments.value = repository.get(id)?.attachments ?: emptyList()
-                    markDirty()
-                }
-            }
-        }
-
-        fun removeAttachment(attachmentId: Long) {
-            viewModelScope.launch {
-                persistence.withLock {
-                    repository.removeAttachment(attachmentId)
-                    val id = loadedId
-                    _attachments.value =
-                        if (id != null) {
-                            repository.get(id)?.attachments ?: emptyList()
-                        } else {
-                            _attachments.value.filterNot { it.id == attachmentId }
-                        }
-                    markDirty()
-                }
-            }
-        }
-
-        private fun tagsForPersistence(): List<String> {
-            val base = _tags.value.filterNot { it == RememberReservedTags.STARRED }
-            return if (_starred.value) {
-                (base + RememberReservedTags.STARRED).distinct()
-            } else {
-                base
-            }
-        }
-
-        private fun currentOptions() =
-            NoteOptions(
-                reminderAt = _reminderAt.value,
-                importance = _importance.value,
-                visibility = _visibility.value,
-                pictureUri = _pictureUri.value,
-                pictureHeroFraming = _pictureHeroFraming.value,
-                locked = _locked.value,
-                iconKey = _iconKey.value,
-                actions = _actions.value,
-                tags = tagsForPersistence(),
-                recurrence = _recurrence.value,
-            )
 
         private fun hasNetChanges(): Boolean {
             val id = loadedId
@@ -442,7 +79,7 @@ class EditNoteViewModel
             val opts = currentOptions()
             val starred = _starred.value
             if (id == null) {
-                return t.isNotBlank() || b.isNotBlank() || _attachments.value.isNotEmpty() || opts.pictureUri != null || opts.tags.isNotEmpty() || opts.reminderAt != null || opts.actions.isNotEmpty() || opts.iconKey != null
+                return t.isNotBlank() || b.isNotBlank() || attachments.value.isNotEmpty() || opts.pictureUri != null || opts.tags.isNotEmpty() || opts.reminderAt != null || opts.actions.isNotEmpty() || opts.iconKey != null
             } else {
                 val old = originalNote ?: return true
                 if (t != old.title) return true
@@ -462,12 +99,7 @@ class EditNoteViewModel
             }
         }
 
-        /**
-         * Persist the in-memory state. Safe to call from ON_STOP, onDispose, or any other
-         * lifecycle hook; concurrent calls serialize through [persistence] so we never
-         * insert the same draft note twice.
-         */
-        suspend fun saveIfNeeded(untitledName: String): (suspend () -> Unit)? {
+        override suspend fun saveIfNeeded(untitledName: String): (suspend () -> Unit)? {
             return persistence.withLock {
                 if (!hasNetChanges()) {
                     persistence.clearDirty()
@@ -544,118 +176,5 @@ class EditNoteViewModel
                     }
                 }
             }
-        }
-
-        private fun syncTimestamps() {
-            val id = loadedId ?: return
-            viewModelScope.launch {
-                val cur = repository.get(id)?.note
-                if (cur != null) {
-                    _createdAt.value = cur.createdAt
-                    _updatedAt.value = cur.updatedAt
-                }
-            }
-        }
-
-        suspend fun trashCurrent() {
-            persistence.withLock {
-                val id = loadedId ?: return@withLock
-                repository.moveToTrash(id)
-                persistence.clearDirty()
-                _trashed.value = true
-                _archived.value = false
-            }
-            syncTimestamps()
-        }
-
-        /**
-         * Flip the note's done state from inside the editor's bottom bar. Routes through
-         * [NoteRepository.markCompleted] / [NoteRepository.markIncomplete] so recurrence
-         * is honored - completing a recurring note rolls it forward instead of moving it
-         * to Done. The live observer in [init] picks up the new value and updates
-         * [completed] automatically; we don't have to mutate it here.
-         */
-        suspend fun toggleCompleted() {
-            val id = loadedId ?: return
-            if (_completed.value) {
-                repository.markIncomplete(id)
-            } else {
-                repository.markCompleted(id)
-            }
-            syncTimestamps()
-        }
-
-        /** Flip the note to the archive shelf after saving any in-flight edits. */
-        suspend fun archiveCurrent(untitledName: String) {
-            // Persist any in-flight edits first so the archive snapshot matches what the user sees.
-            saveIfNeeded(untitledName)
-            persistence.withLock {
-                val id = loadedId ?: return@withLock
-                repository.archiveNote(id)
-                _archived.value = true
-                _trashed.value = false
-                persistence.clearDirty()
-            }
-            syncTimestamps()
-        }
-
-        suspend fun unarchiveCurrent() {
-            persistence.withLock {
-                val id = loadedId ?: return@withLock
-                repository.unarchiveNote(id)
-                _archived.value = false
-                _trashed.value = false
-                persistence.clearDirty()
-            }
-            syncTimestamps()
-        }
-
-        suspend fun restoreFromTrashCurrent() {
-            persistence.withLock {
-                val id = loadedId ?: return@withLock
-                repository.restoreFromTrash(id)
-                _trashed.value = false
-                _archived.value = false
-                persistence.clearDirty()
-            }
-            syncTimestamps()
-        }
-
-        /**
-         * Permanent delete of the current note. Called from the trashed-state action bar
-         * after the user confirms the AlertDialog in [EditNoteScreen]. Mirrors
-         * [EditListViewModel.deleteForeverCurrent]; without this method the call site
-         * was unresolved.
-         */
-        suspend fun deleteForeverCurrent() {
-            persistence.withLock {
-                val id = loadedId ?: return@withLock
-                repository.deleteForever(id)
-                persistence.clearDirty()
-            }
-        }
-
-        suspend fun fireNotification(
-            context: android.content.Context,
-            untitledName: String,
-        ) {
-            saveIfNeeded(untitledName)
-            val id = loadedId ?: return
-            val note = repository.get(id)?.note ?: return
-            val reminderPrefs =
-                EntryPointAccessors
-                    .fromApplication(
-                        context.applicationContext,
-                        SettingsDependenciesEntryPoint::class.java,
-                    ).reminderPrefs()
-            val keepUntilDone = reminderPrefs.snapshot().keepReminderNotificationsUntilDone
-            dev.bikram.remember.reminders.ReminderReceiver
-                .showNotification(context, note, keepUntilDone = keepUntilDone)
-            android.widget.Toast
-                .makeText(
-                    context,
-                    context.getString(dev.bikram.remember.R.string.notification_created),
-                    android.widget.Toast.LENGTH_SHORT,
-                ).show()
         }
     }
