+++ app/src/main/java/dev/bikram/remember/ui/onboarding/OnboardingPermissionsScreen.kt
@@ -21,7 +21,6 @@ import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
@@ -77,6 +76,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import dev.bikram.remember.R
+import dev.bikram.remember.ui.common.isSmallLandscape
 import dev.bikram.remember.ui.common.RememberMaterialRoundedSymbol
 import dev.bikram.remember.ui.components.RememberButton
 import dev.bikram.remember.ui.components.RememberTextButton
@@ -125,12 +125,12 @@ fun OnboardingPermissionsScreen(
         modifier = Modifier.fillMaxSize(),
         color = scheme.background,
     ) {
-        BoxWithConstraints(
+        val useTwoColumns = isSmallLandscape()
+        Box(
             modifier =
                 Modifier
                     .fillMaxSize(),
         ) {
-            val useTwoColumns = maxHeight < 480.dp && maxWidth > maxHeight
             if (useTwoColumns) {
                 Row(
                     modifier =
